feat:标签增加分类功能

+++ cqupt-hub-subject/cqupt-hub-infra/src/main/java/com/lb/subject/infra/basic/entity/SubjectCategory.java
@@ -23,7 +23,7 @@ public class SubjectCategory implements Serializable {
     /**
      * 分类类型
      */
-    private Byte categoryType;
+    private Integer categoryType;
 
     /**
      * 图标连接
